Added files utils module and migrated relevant code there
